refactor for demo
