+++ protocol/deployments.json
@@ -1,6 +1,1 @@
-{
-  "chain-31337": {
-    "MultiplierVerifier": "0x5FbDB2315678afecb367f032d93F642f64180aa3",
-    "MultiAssetShieldedPool": "0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512"
-  }
-}
+{}
