+++ protocol/deployments/localhost/.chainId
@@ -1,0 +1,1 @@
+31337
