+++ protocol/src/log.ts
@@ -1,0 +1,7 @@
+export function logAction(str: string) {
+  console.log("-----------------------------------------------");
+  console.log(" " + str);
+  console.log("-----------------------------------------------");
+}
+
+
